Tidy up quick fix infra to be reusable across linters
