+++ src/diagnostics/linter.ts
@@ -1,7 +1,8 @@
-import { TextDocument, Diagnostic } from "vscode";
+import { TextDocument, Diagnostic, CodeAction } from "vscode";
 
 import * as ast from '../porter/ast';
 
 export interface Linter {
     lint(document: TextDocument, manifest: ast.PorterManifestYAML): Promise<Diagnostic[]>;
+    fixes(document: TextDocument, manifest: ast.PorterManifestYAML, diagnostics: Diagnostic[]): CodeAction[];
 }
